+++ CHANGELOG.md
@@ -11,11 +11,13 @@ for further information about branching and tagging conventions.
 #### Added
 - Atomic swap requests history in atomic swap details
 - Update asset form
+- Mass payment operation
 - Set default quote asset for business in settings
 - Convert asset for each business
 
 #### Deleted
 - Dropdown account balances
+- Issue form
 
 ## [1.0.0-rc.6] - 2019-08-02
 #### Fixed